# docs/api/openapi.yml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Whatboo API",
    "version": "2026.06.30",
    "description": "Contrato operativo para la API multi-tenant de Whatboo. Todas las rutas privadas usan el prefijo global /api/v1, JWT bearer y cabecera tenant salvo que se indique lo contrario."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API base"
    }
  ],
...
  "security": [
    {
      "bearerAuth": [],
      "tenantSlug": []
    }
  ],
  "paths": {
...
    "/reservations/today": {
      "get": {
        "tags": ["Reservations"],
        "summary": "Reservas del dia actual",
        "parameters": [{ "$ref": "#/components/parameters/TenantSlugHeader" }],
        "responses": {
          "200": {
            "description": "Reservas del dia",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Reservation" }
                }
              }
            }
          }
        }
      }
    },
    "/reservations": {
      "get": {
        "tags": ["Reservations"],
        "summary": "Listar reservas con filtros dashboard",
        "parameters": [
          { "$ref": "#/components/parameters/TenantSlugHeader" },
          { "name": "date", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "from", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "to", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "status", "in": "query", "schema": { "$ref": "#/components/schemas/ReservationStatus" } },
          { "name": "source", "in": "query", "schema": { "$ref": "#/components/schemas/ReservationSource" } },
          { "$ref": "#/components/parameters/SearchQuery" },
          { "$ref": "#/components/parameters/TakeQuery" },
          { "$ref": "#/components/parameters/SkipQuery" }
        ],
        "responses": {
          "200": {
            "description": "Reservas paginadas o lista diaria",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    { "$ref": "#/components/schemas/PaginatedReservations" },
                    {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Reservation" }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Reservations"],
        "summary": "Crear reserva manual",
        "parameters": [{ "$ref": "#/components/parameters/TenantSlugHeader" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateReservationRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reserva creada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Reservation" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/reservations/{id}": {
      "get": {
        "tags": ["Reservations"],
        "summary": "Ver reserva",
        "parameters": [
          { "$ref": "#/components/parameters/TenantSlugHeader" },
          { "$ref": "#/components/parameters/IdPath" }
        ],
        "responses": {
          "200": {
            "description": "Reserva",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Reservation" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Reservations"],
        "summary": "Actualizar fecha, mesa, pax o notas con revalidacion de disponibilidad",
        "parameters": [
          { "$ref": "#/components/parameters/TenantSlugHeader" },
          { "$ref": "#/components/parameters/IdPath" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateReservationRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reserva actualizada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Reservation" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/reservations/{id}/status": {
      "patch": {
        "tags": ["Reservations"],
        "summary": "Actualizar estado excepto cancelacion y no-show",
        "parameters": [
          { "$ref": "#/components/parameters/TenantSlugHeader" },
          { "$ref": "#/components/parameters/IdPath" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateReservationStatusRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Estado actualizado" },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/reservations/{id}/cancel": {
      "patch": {
        "tags": ["Reservations"],
        "summary": "Cancelar reserva con motivo estructurado",
        "parameters": [
          { "$ref": "#/components/parameters/TenantSlugHeader" },
          { "$ref": "#/components/parameters/IdPath" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CancelReservationRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Reserva cancelada" },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/reservations/{id}/no-show": {
      "patch": {
        "tags": ["Reservations"],
        "summary": "Marcar reserva como no-show con motivo estructurado",
        "parameters": [
          { "$ref": "#/components/parameters/TenantSlugHeader" },
          { "$ref": "#/components/parameters/IdPath" }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NoShowReservationRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Reserva marcada como no-show" },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/reservations/{id}/cancellation-audits": {
      "get": {
        "tags": ["Reservations"],
        "summary": "Listar auditoria de cancelacion de una reserva",
        "parameters": [
          { "$ref": "#/components/parameters/TenantSlugHeader" },
          { "$ref": "#/components/parameters/IdPath" }
        ],
        "responses": {
          "200": {
            "description": "Eventos de cancelacion",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ReservationCancellationAudit" }
                }
              }
            }
          }
        }
      }
    },
    "/reservations/{id}/no-show-audits": {
      "get": {
        "tags": ["Reservations"],
        "summary": "Listar auditoria estructurada de no-show de una reserva",
        "parameters": [
          { "$ref": "#/components/parameters/TenantSlugHeader" },
          { "$ref": "#/components/parameters/IdPath" }
        ],
        "responses": {
          "200": {
            "description": "Eventos de no-show",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ReservationNoShowAudit" }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "tenantSlug": { "type": "apiKey", "in": "header", "name": "X-Tenant-Slug" },
      "onboardingToken": { "type": "apiKey", "in": "header", "name": "X-Onboarding-Token" }
    },
    "parameters": {
      "TenantSlugHeader": {
        "name": "X-Tenant-Slug",
        "in": "header",
        "required": false,
        "schema": { "type": "string" }
      },
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
      "SearchQuery": {
        "name": "q",
        "in": "query",
        "required": false,
        "schema": { "type": "string" }
      },
      "TakeQuery": {
        "name": "take",
        "in": "query",
        "required": false,
        "schema": { "type": "integer", "minimum": 1, "maximum": 100 }
      },
      "SkipQuery": {
        "name": "skip",
        "in": "query",
        "required": false,
        "schema": { "type": "integer", "minimum": 0 }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Solicitud invalida",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
    },
    "schemas": {
...
      "ReservationStatus": {
        "type": "string",
        "enum": ["PENDING", "CONFIRMED", "CANCELLED", "COMPLETED", "NO_SHOW"]
      },
      "ReservationSource": {
        "type": "string",
        "enum": ["WHATSAPP", "DASHBOARD", "PHONE", "WALK_IN", "API"]
      },
...
      "Customer": {
        "type": "object",
        "required": ["id", "restaurantId", "phone", "active"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "restaurantId": { "type": "string", "format": "uuid" },
          "firstName": { "type": ["string", "null"] },
          "lastName": { "type": ["string", "null"] },
          "phone": { "type": "string" },
          "email": { "type": ["string", "null"], "format": "email" },
          "preferredLanguage": { "type": "string" },
          "notes": { "type": ["string", "null"] },
          "totalReservations": { "type": "integer" },
          "lastReservationAt": { "type": ["string", "null"], "format": "date-time" },
          "active": { "type": "boolean" }
        }
      },
...
      "Reservation": {
        "type": "object",
        "required": ["id", "restaurantId", "customerId", "reservationStart", "reservationEnd", "guestCount", "status"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "restaurantId": { "type": "string", "format": "uuid" },
          "customerId": { "type": "string", "format": "uuid" },
          "reservationDate": { "type": "string", "format": "date-time" },
          "reservationStart": { "type": "string", "format": "date-time" },
          "reservationEnd": { "type": "string", "format": "date-time" },
          "guestCount": { "type": "integer", "minimum": 1 },
          "status": { "$ref": "#/components/schemas/ReservationStatus" },
          "source": { "$ref": "#/components/schemas/ReservationSource" },
          "notes": { "type": ["string", "null"] },
          "confirmationCode": { "type": ["string", "null"] },
          "customer": { "$ref": "#/components/schemas/Customer" },
          "assignedTables": { "type": "array", "items": { "type": "object" } }
        }
      },
      "CreateReservationRequest": {
        "type": "object",
        "required": ["phone", "guestCount", "reservationStart"],
        "properties": {
          "phone": { "type": "string" },
          "customerName": { "type": "string" },
          "guestCount": { "type": "integer", "minimum": 1 },
          "reservationStart": { "type": "string", "format": "date-time" },
          "durationMinutes": { "type": "integer", "minimum": 15 },
          "notes": { "type": "string" }
        }
      },
      "UpdateReservationRequest": {
        "type": "object",
        "properties": {
          "reservationStart": { "type": "string", "format": "date-time" },
          "reservationEnd": { "type": "string", "format": "date-time" },
          "guestCount": { "type": "integer", "minimum": 1 },
          "notes": { "type": ["string", "null"] },
          "tableId": { "type": ["string", "null"], "format": "uuid" }
        }
      },
      "UpdateReservationStatusRequest": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": { "$ref": "#/components/schemas/ReservationStatus" }
        }
      },
      "CancelReservationRequest": {
        "type": "object",
        "properties": {
          "reasonCode": {
            "type": "string",
            "enum": ["CUSTOMER_REQUEST", "RESTAURANT_CLOSED", "DUPLICATE_BOOKING", "PAYMENT_OR_DEPOSIT_ISSUE", "OPERATIONAL_CAPACITY", "OTHER"]
          },
          "details": { "type": "string", "maxLength": 500 },
          "source": { "$ref": "#/components/schemas/ReservationSource" }
        }
      },
      "NoShowReservationRequest": {
        "type": "object",
        "properties": {
          "reasonCode": {
            "type": "string",
            "enum": ["CUSTOMER_DID_NOT_ARRIVE", "CUSTOMER_UNREACHABLE", "ARRIVED_TOO_LATE", "DUPLICATE_BOOKING", "OTHER"]
          },
          "details": { "type": "string", "maxLength": 500 }
        }
      },
      "ReservationCancellationAudit": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "reservationId": { "type": "string", "format": "uuid" },
          "reason": { "type": "string" },
          "source": { "$ref": "#/components/schemas/ReservationSource" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "ReservationNoShowAudit": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "reservationId": { "type": "string", "format": "uuid" },
          "reasonCode": {
            "type": "string",
            "enum": ["CUSTOMER_DID_NOT_ARRIVE", "CUSTOMER_UNREACHABLE", "ARRIVED_TOO_LATE", "DUPLICATE_BOOKING", "OTHER"]
          },
          "details": { "type": ["string", "null"] },
          "source": { "$ref": "#/components/schemas/ReservationSource" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "PaginatedReservations": {
        "type": "object",
        "required": ["data", "total", "take", "skip"],
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Reservation" } },
          "total": { "type": "integer" },
          "take": { "type": "integer" },
          "skip": { "type": "integer" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "integer" },
          "message": {
            "oneOf": [
              { "type": "string" },
              { "type": "array", "items": { "type": "string" } }
            ]
          },
          "error": { "type": "string" }
        }
      }
    }
  }
}